move split_polygon_on_repeated_edges to intersection module
