+++ crates/bermuda/tests/test_intersection.rs
@@ -276,3 +276,20 @@ fn test_find_intersection_points_cross_intersect_in_point() {
 
     assert_eq!(result, expected);
 }
+
+#[rstest]
+#[case::rectangle(vec![Point::new(0.0, 0.0), Point::new(1.0, 0.0), Point::new(1.0, 1.0), Point::new(0.0, 1.0)], 1, 4)]
+#[case::rectangle_in_rectangle(vec![Point::new(0.0, 0.0), Point::new(3.0, 0.0), Point::new(3.0, 3.0), Point::new(0.0, 3.0), Point::new(0.0, 0.0), Point::new(1.0, 1.0),
+    Point::new(2.0, 1.0),
+    Point::new(2.0, 2.0),
+    Point::new(1.0, 2.0),
+    Point::new(1.0, 1.0),], 2, 8)]
+fn test_split_polygon_on_repeated_edges(
+    #[case] polygon: Vec<Point>,
+    #[case] polygon_count: usize,
+    #[case] edge_count: usize,
+) {
+    let (polygons, edges) = intersection::split_polygon_on_repeated_edges(&polygon);
+    assert_eq!(polygons.len(), polygon_count);
+    assert_eq!(edges.len(), edge_count);
+}
