+++ crates/triangulation/src/intersection.rs
@@ -510,7 +510,7 @@ pub fn find_intersection_points(polygon_list: &[Vec<point::Point>]) -> Vec<Vec<p
             }
             Intersection::CollinearWithOverlap((p1, p2)) => {
                 // Handle collinear overlap case if needed
-                // Add both points to both edges
+                // to add both points to both edges
                 for point in [p1, p2] {
                     intersections_points
                         .entry(intersection.first())
@@ -581,4 +581,134 @@ pub fn find_intersection_points(polygon_list: &[Vec<point::Point>]) -> Vec<Vec<p
     }
 
     new_polygons_list
+}
+
+struct GraphNode {
+    edges: Vec<point::Point>,
+    visited: bool,
+    sub_index: usize,
+}
+
+impl Default for GraphNode {
+    fn default() -> Self {
+        GraphNode {
+            edges: Vec::new(),
+            visited: false,
+            sub_index: 0,
+        }
+    }
+}
+
+/// Splits a polygon into multiple polygons based on repeated edges using a DFS graph traversal.
+///
+/// This function identifies duplicated edges within the polygon and splits it into multiple
+/// disjoint polygonal components. The splitting is guided by the unique edges that are determined
+/// through a deduplication process. The function returns the resulting polygons and the list of
+/// deduplicated edges.
+///
+/// # Arguments
+/// * `polygon` - A slice of `Point`s representing a polygon to be split.
+///
+/// # Returns
+/// A tuple containing:
+/// * A `Vec<Vec<Point>>` representing the split polygons as individual vectors of points.
+/// * A `Vec<Segment>` containing the deduplicated list of edges used during splitting.
+///
+/// # Purpose
+/// This function is designed for edge triangulation (from `path_triangulation.rs`) and
+/// provides the edges that can be further used for face triangulation.
+///
+/// # Examples
+/// ```
+/// use triangulation::point::{Point, Segment};
+/// use triangulation::intersection::split_polygon_on_repeated_edges;
+///
+/// let polygon = vec![
+///     Point::new(0.0, 0.0),
+///     Point::new(3.0, 0.0),
+///     Point::new(3.0, 3.0),
+///     Point::new(0.0, 3.0),
+///     Point::new(0.0, 0.0),
+///     Point::new(1.0, 1.0),
+///     Point::new(2.0, 1.0),
+///     Point::new(2.0, 2.0),
+///     Point::new(1.0, 2.0),
+///     Point::new(1.0, 1.0),
+/// ];
+///
+/// let (polygons, edges) = split_polygon_on_repeated_edges(&polygon);
+///
+/// assert_eq!(polygons.len(), 2); // The polygon is split into two disjoint polygons.
+/// assert_eq!(edges.len(), 8); // Deduplicated edges used for splitting.
+/// ```
+#[inline]
+pub fn split_polygon_on_repeated_edges(
+    polygon: &[point::Point],
+) -> (Vec<Vec<point::Point>>, Vec<point::Segment>) {
+    let edges_dedup = point::calc_dedup_edges(&[polygon.to_vec()]);
+    let mut result = Vec::new();
+
+    // Convert deduped edges to HashSet for efficient lookup
+    let edges_set: HashSet<point::Segment> = edges_dedup.clone().into_iter().collect();
+    let mut edges_map: HashMap<point::Point, GraphNode> = HashMap::new();
+
+    // Build edges map from consecutive points
+    for window in polygon.windows(2) {
+        let segment = point::Segment::new(window[0], window[1]);
+        if edges_set.contains(&segment) {
+            edges_map
+                .entry(window[0])
+                .or_default()
+                .edges
+                .push(window[1]);
+        }
+    }
+
+    // Handle the edge between last and first point
+    if let (Some(back), Some(front)) = (polygon.last(), polygon.first()) {
+        let segment = point::Segment::new(*back, *front);
+        if edges_set.contains(&segment) {
+            edges_map.entry(*back).or_default().edges.push(*front);
+        }
+    }
+
+    // Process all edges
+    let mut edges_to_process: Vec<_> = edges_map.keys().cloned().collect();
+    while let Some(start_point) = edges_to_process.pop() {
+        if let Some(edge) = edges_map.get_mut(&start_point) {
+            if edge.visited {
+                continue;
+            }
+
+            edge.visited = true;
+            let mut new_polygon = vec![start_point];
+            let mut current_point = start_point;
+
+            // Follow the edge chain
+            while let Some(node) = edges_map.get_mut(&current_point) {
+                if node.sub_index >= node.edges.len() {
+                    break;
+                }
+
+                let next_point = node.edges[node.sub_index];
+                node.sub_index += 1;
+
+                if let Some(next_node) = edges_map.get_mut(&next_point) {
+                    next_node.visited = true;
+                }
+
+                new_polygon.push(next_point);
+                current_point = next_point;
+            }
+
+            // Remove duplicate points at the start/end
+            while new_polygon.len() > 1 && new_polygon.first() == new_polygon.last() {
+                new_polygon.pop();
+            }
+
+            result.push(new_polygon);
+        }
+    }
+
+    (result, edges_dedup)
 }
